requirements.txt
geopandas
python-dotenv
psycopg2-binary
SQLAlchemy<2.0
